AddLesson07AndUpdateLesson2AndUpDateHelpers

+++ Lesson04Task01/Class/Employee.cs
@@ -10,7 +10,7 @@ namespace Task01.Class
     {
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public int Age { get; set; }
+        public byte Age { get; set; }
         public char Sex { get; set; }
         public string PESEL { get; set; }
         public long Id { get; set; }
